Moved tasks_crush.py and tasks_oogway.py into directory robot_tasks for better organization and easier use
