+++ onboard/src/task_planning/task_planning/task_runner.py
@@ -8,7 +8,7 @@ from rclpy.clock import Clock
 from rclpy.duration import Duration
 from rclpy.node import Node
 
-from task_planning import tasks_crush, tasks_oogway
+from robot_tasks import tasks_crush, tasks_oogway
 from task_planning.interface.controls import Controls
 from task_planning.interface.cv import CV
 from task_planning.interface.marker_dropper import MarkerDropper
